% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baca_data.R
\name{baca_data}
\alias{baca_data}
\title{Baca Data Internal Green Charge}
\usage{
baca_data()
}
\value{
List berisi dataframe stasiun, transaksi, konsumsi_harian, historis
}
\description{
Baca Data Internal Green Charge
}
