% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulan_historis}
\alias{bulan_historis}
\title{Data Historis Bulanan}
\format{
Data frame dengan kolom:
\describe{
\item{bulan}{Nama bulan}
\item{periode}{Tanggal awal bulan}
\item{total_kwh}{Total energi tersalurkan dalam kWh}
\item{total_sesi}{Total sesi pengisian}
\item{total_pendapatan}{Total pendapatan dalam rupiah}
}
}
\usage{
bulan_historis
}
\description{
Data agregasi historis bulanan seluruh stasiun
periode Januari-Juni 2026.
}
\keyword{datasets}
