% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteksi_anomali.R
\name{deteksi_anomali}
\alias{deteksi_anomali}
\title{Deteksi Anomali Konsumsi Energi}
\usage{
deteksi_anomali(konsumsi_harian)
}
\arguments{
\item{konsumsi_harian}{Dataframe konsumsi harian}
}
\value{
Dataframe baris-baris yang terdeteksi anomali
}
\description{
Memberi peringatan jika konsumsi harian stasiun berada di luar
rentang rata-rata +/- 2 standar deviasi
}
