% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_stasiun.R
\name{kpi_stasiun}
\alias{kpi_stasiun}
\title{KPI Lengkap per Stasiun}
\usage{
kpi_stasiun(transaksi, konsumsi_harian)
}
\arguments{
\item{transaksi}{Dataframe transaksi}

\item{konsumsi_harian}{Dataframe konsumsi harian}
}
\value{
Dataframe KPI per stasiun
}
\description{
KPI Lengkap per Stasiun
}
