% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_konsumsi.R
\name{plot_konsumsi}
\alias{plot_konsumsi}
\title{Visualisasi Tren Konsumsi Energi Harian}
\usage{
plot_konsumsi(konsumsi_harian)
}
\arguments{
\item{konsumsi_harian}{Dataframe konsumsi harian}
}
\value{
ggplot object
}
\description{
Visualisasi Tren Konsumsi Energi Harian
}
