% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringkasan_energi.R
\name{ringkasan_energi}
\alias{ringkasan_energi}
\title{Ringkasan Konsumsi Energi per Stasiun}
\usage{
ringkasan_energi(transaksi)
}
\arguments{
\item{transaksi}{Dataframe transaksi}
}
\value{
Dataframe ringkasan energi per stasiun
}
\description{
Ringkasan Konsumsi Energi per Stasiun
}
