% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stasiun}
\alias{stasiun}
\title{Data Stasiun Pengisian}
\format{
Data frame dengan kolom:
\describe{
\item{id_stasiun}{ID unik stasiun}
\item{nama_stasiun}{Nama stasiun pengisian}
\item{kota}{Kota lokasi stasiun}
\item{jumlah_charger}{Jumlah charger tersedia}
\item{tarif_per_kwh}{Tarif per kWh dalam rupiah}
}
}
\usage{
stasiun
}
\description{
Data informasi 10 stasiun pengisian kendaraan listrik
PT Green Charge Indonesia.
}
\keyword{datasets}
