% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baca_data.R
\name{baca_data}
\alias{baca_data}
\title{Baca Data CSV Green Charge}
\usage{
baca_data(path)
}
\arguments{
\item{path}{Path folder tempat file CSV disimpan}
}
\value{
List berisi dataframe stasiun, transaksi, konsumsi_harian, historis
}
\description{
Baca Data CSV Green Charge
}
